Time: 0 ms (100.00%) | Memory: 10.3 MB (33.86%) - LeetSync

+++ 1878-check-if-array-is-sorted-and-rotated/check-if-array-is-sorted-and-rotated.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool check(vector<int>& nums) {
         int count=0;
-        long n=nums.size();
+        int n=nums.size();
         for(int i=0;i<n;i++){
             if(nums[i]>nums[(i+1)%n]){
                 count++;
